refactoring Transformers depending on statement types
