+++ src/main/java/org/stdg/SqlQuery.java
@@ -13,12 +13,6 @@
 
 package org.stdg;
 
-import net.sf.jsqlparser.JSQLParserException;
-import net.sf.jsqlparser.parser.CCJSqlParserUtil;
-import net.sf.jsqlparser.statement.Statement;
-import net.sf.jsqlparser.statement.select.Select;
-import net.sf.jsqlparser.statement.update.Update;
-
 import java.util.*;
 
 import static java.util.stream.Collectors.joining;
@@ -37,30 +31,6 @@ public class SqlQuery {
     public SqlQuery(String queryAsString, List<Object> parameters) {
         this.queryAsString = queryAsString;
         this.parameters = parameters;
-    }
-
-    Optional<SqlQuery> transformToSelectQuery() {
-        Statement parsedStatement = parse(queryAsString);
-        if(parsedStatement instanceof Select) {
-            return Optional.of(this);
-        }
-        if(parsedStatement instanceof Update) {
-            Update update = (Update) parsedStatement;
-            SelectTransformer selectTransformer = new SelectTransformer(update);
-            String selectQueryAsString = selectTransformer.transformToSelect();
-            SqlQuery sqlQuery = new SqlQuery(selectQueryAsString, parameters);
-            return Optional.of(sqlQuery);
-        }
-        return Optional.empty();
-    }
-
-   private Statement parse(String sqlQuery) {
-        try {
-            return CCJSqlParserUtil.parse(sqlQuery);
-        } catch (JSQLParserException e) {
-            e.printStackTrace();
-            return null;
-        }
     }
 
     static SqlQuery buildFromRow(DatasetRow rowToSearch) {
