+++ src/test/java/org/stdg/test/H2Configuration.java
@@ -13,7 +13,6 @@
 
 package org.stdg.test;
 
-import net.ttddyy.dsproxy.support.ProxyDataSource;
 import org.junit.jupiter.api.BeforeAll;
 import org.quickperf.junit5.QuickPerfTest;
